FIX : ts refactor
