+++ src/components/Inputs/AuthTextAreaWithLabel/style.ts
@@ -1,0 +1,1 @@
+export {};
