+++ src/pages/TimeTablePage/style.ts
@@ -1,0 +1,1 @@
+export {};
